third_party/blink/web_tests/external/wpt/html/rendering/non-replaced-elements/the-fieldset-and-legend-elements/legend-display-ruby-expected.txt

This is a testharness.js-based test.
[FAIL] rendered legend with display: ruby
  assert_equals: display: ruby is not supported expected "ruby" but got "block ruby"
[FAIL] rendered legend with display: ruby-base
  assert_equals: display: ruby-base is not supported expected "ruby-base" but got "block"
[FAIL] rendered legend with display: ruby-text
  assert_equals: display: ruby-text is not supported expected "ruby-text" but got "block"
[FAIL] rendered legend with display: ruby-base-container
  assert_equals: display: ruby-base-container is not supported expected "ruby-base-container" but got "block"
[FAIL] rendered legend with display: ruby-text-container
  assert_equals: display: ruby-text-container is not supported expected "ruby-text-container" but got "block"
[FAIL] rendered legend with display: ruby; overflow:hidden
  assert_equals: display: ruby is not supported expected "ruby" but got "block ruby"
[FAIL] rendered legend with display: ruby-base; overflow:hidden
  assert_equals: display: ruby-base is not supported expected "ruby-base" but got "block"
[FAIL] rendered legend with display: ruby-text; overflow:hidden
  assert_equals: display: ruby-text is not supported expected "ruby-text" but got "block"
[FAIL] rendered legend with display: ruby-base-container; overflow:hidden
  assert_equals: display: ruby-base-container is not supported expected "ruby-base-container" but got "block"
[FAIL] rendered legend with display: ruby-text-container; overflow:hidden
  assert_equals: display: ruby-text-container is not supported expected "ruby-text-container" but got "block"
[FAIL] rendered legend with display: ruby; columns:1
  assert_equals: display: ruby is not supported expected "ruby" but got "block ruby"
[FAIL] rendered legend with display: ruby-base; columns:1
  assert_equals: display: ruby-base is not supported expected "ruby-base" but got "block"
[FAIL] rendered legend with display: ruby-text; columns:1
  assert_equals: display: ruby-text is not supported expected "ruby-text" but got "block"
[FAIL] rendered legend with display: ruby-base-container; columns:1
  assert_equals: display: ruby-base-container is not supported expected "ruby-base-container" but got "block"
[FAIL] rendered legend with display: ruby-text-container; columns:1
  assert_equals: display: ruby-text-container is not supported expected "ruby-text-container" but got "block"
[FAIL] rendered legend with display: ruby; overflow:hidden;columns:1
  assert_equals: display: ruby is not supported expected "ruby" but got "block ruby"
[FAIL] rendered legend with display: ruby-base; overflow:hidden;columns:1
  assert_equals: display: ruby-base is not supported expected "ruby-base" but got "block"
[FAIL] rendered legend with display: ruby-text; overflow:hidden;columns:1
  assert_equals: display: ruby-text is not supported expected "ruby-text" but got "block"
[FAIL] rendered legend with display: ruby-base-container; overflow:hidden;columns:1
  assert_equals: display: ruby-base-container is not supported expected "ruby-base-container" but got "block"
[FAIL] rendered legend with display: ruby-text-container; overflow:hidden;columns:1
  assert_equals: display: ruby-text-container is not supported expected "ruby-text-container" but got "block"
Harness: the test ran to completion.
